Added employee details in security context, removed employee service, and, related changes and other changes
