+++ src/main/java/com/sba6/srm/controller/RequestController.java
@@ -59,8 +59,6 @@ public class RequestController {
 	private CommentService commentService;
 	@Autowired
 	private ReasonService reasonService;
-	@Autowired
-	private AuthenticationController authenticationController;
 	
 	
 	//1. GET Requests for manager GET/api/requests/{mgrId}
@@ -111,12 +109,6 @@ public class RequestController {
 		emailService.mailAddRequest(request, request.getEmployee().getEmail());
 		return new ResponseEntity(HttpStatus.OK);
 	}
-
-	@ApiOperation(value = "Get employee details")
-	@GetMapping(value = "/api/emp/{empId}")
-	public Employee getEmployee(@PathVariable Long empId) {
-		return employeeService.getEmployee(empId);
-	}
 	
 	//6. Add request comment POST/api/request/comment
 	@ApiOperation(value = "Add request comment")
@@ -133,42 +125,4 @@ public class RequestController {
 		return commentService.getCommentsForRequest(requestId);
 	}
 	
-	
-	@PostMapping(value="/api/google")
-	public ResponseEntity googleLogin(@RequestBody String googleCode){
-		
-		//1. Get tokens by exchanging the google code
-		final String postUrl = "https://accounts.google.com/o/oauth2/token";
-		RestTemplate restTemplate = new RestTemplate();
-		String input = "code="+googleCode+"&"
-		+"client_id=698682231712-gbno37u3fdovq1rjrhurj4o6bo00okg9.apps.googleusercontent.com&"
-		+"client_secret=IImw1dZ0P1tm2tq82QeXPW19&"
-		+"redirect_uri=http://localhost:4200/login&"
-		+"grant_type=authorization_code";
-		HttpHeaders headers = new HttpHeaders();
-		headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
-		HttpEntity<String> request = new HttpEntity<String>(input, headers);
-		ResponseEntity<GoogleTokens> tokensResponse = restTemplate.postForEntity(postUrl, request, GoogleTokens.class);
-			
-		//2. Get email by sending access token
-		String accessToken = tokensResponse.getBody().getAccess_token();
-		String getUrl = "https://www.googleapis.com/oauth2/v2/userinfo";
-		restTemplate = new RestTemplate();
-		headers = new HttpHeaders();
-		String authorizationInput = "Bearer "+accessToken;
-		headers.set("Authorization", authorizationInput);
-		request = new HttpEntity<>("parameters", headers);
-		ResponseEntity<GoogleResponse> googleResponse = restTemplate.exchange(getUrl, HttpMethod.GET, request, GoogleResponse.class);
-		
-		//3. Verify email in db
-		Employee theEmployee = employeeService.getEmployee(googleResponse.getBody().getEmail());
-		if(theEmployee == null) {
-			 return ResponseEntity.ok(null);
-		}
-		LoginDetail loginDetail = new LoginDetail();
-		loginDetail.setUserName(theEmployee.getLoginDetail().getUserName());
-		loginDetail.setPassword(theEmployee.getLoginDetail().getPassword());
-		return authenticationController.authenticateUser(loginDetail);
-
-	}
 }
