+++ src/main/web-app/src/app/base-view/base-view.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthService } from '../services/auth.service';
-import { EmployeeService } from '../services/employee.service';
 
 @Component({
   selector: 'app-base-view',
@@ -10,8 +9,7 @@ import { EmployeeService } from '../services/employee.service';
 export class BaseViewComponent implements OnInit {
 
   constructor(
-    public authService: AuthService,
-    public employeeService: EmployeeService
+    public authService: AuthService
   ) {
     if (authService.securityContext.role === 'EMPLOYEE') {
       this.navLinks.shift();
