+++ src/main/web-app/src/app/models/security-context.ts
@@ -1,0 +1,11 @@
+export class SecurityContext {
+	id: number;
+	email: string;
+	firstName: string;
+	lastName: string;
+	role: string;
+	token: string;
+	designation: string;
+	department: string;
+	managerName: string;
+}
