--- src/main/web-app/src/app/security-context.ts
@@ -1,8 +1,0 @@
-export class SecurityContext {
-    id:number;
-	role:string;
-	firstName:string;
-	lastName:string;
-	email:string;
-	token:string;
-}
