--- src/main/web-app/src/app/services/employee.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { EmployeeService } from './employee.service';
-
-describe('EmployeeService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: EmployeeService = TestBed.get(EmployeeService);
-    expect(service).toBeTruthy();
-  });
-});
