--- src/main/web-app/src/app/services/employee.service.ts
@@ -1,21 +1,0 @@
-import { Injectable } from '@angular/core';
-import { Observable, of } from 'rxjs';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { catchError, map, tap } from 'rxjs/operators';
-import { Employee } from '../models/employee';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class EmployeeService {
-
-  private requestEmployeeUrl = '/api/emp/';
-
-  constructor(private http: HttpClient) { }
-
-  getEmployee(empId: number): Observable<Employee> {
-    return this.http.get<Employee>(this.requestEmployeeUrl + `${empId}`);
-  }
-
-
-}
